refac: datalistcomp style 분리
