+++ test-web/src/components/DataListView/DataListComp.js
@@ -4,6 +4,7 @@ import Spinner from 'react-bootstrap/Spinner';
 import DataList from './DataList';
 import Pagination from './Pagination';
 import { useMeatList } from '../../API/get/getMeatListSWR';
+import style from './style/datalistcompstyle';
 
 // 데이터 목록 조회 페이지 컴포넌트
 const DataListComp = ({
@@ -136,29 +137,3 @@ const DataListComp = ({
 };
 
 export default DataListComp;
-
-const style = {
-  listContainer: {
-    textAlign: 'center',
-    width: '100%',
-    paddingRight: '0px',
-    paddingBottom: '0',
-    height: 'auto',
-    minWidth: '640px',
-  },
-  paginationBar: {
-    marginTop: '40px',
-    width: '100%',
-    justifyContent: 'center',
-  },
-  paginationContainer: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    minWidth: '640px',
-  },
-  formControl: {
-    minWidth: 120,
-    marginLeft: '20px',
-  },
-};
