+++ test-web/src/components/DataListView/style/datalistcompstyle.js
@@ -1,0 +1,27 @@
+const style = {
+    listContainer: {
+      textAlign: 'center',
+      width: '100%',
+      paddingRight: '0px',
+      paddingBottom: '0',
+      height: 'auto',
+      minWidth: '640px',
+    },
+    paginationBar: {
+      marginTop: '40px',
+      width: '100%',
+      justifyContent: 'center',
+    },
+    paginationContainer: {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+      minWidth: '640px',
+    },
+    formControl: {
+      minWidth: 120,
+      marginLeft: '20px',
+    },
+  };
+
+  export default style
